完整功能版本

+++ BackEnd/ProcessManagement_Microservice/src/main/java/com/example/processmanagement_microservice/service/SponsorshipService.java
@@ -4,12 +4,6 @@ import com.example.processmanagement_microservice.model.Sponsorship;
 import com.example.processmanagement_microservice.dao.SponsorshipDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
-import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.query.Criteria;
-import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.data.mongodb.core.query.Update;
-import com.mongodb.client.result.DeleteResult;
-import com.mongodb.client.result.UpdateResult;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,6 +13,16 @@ import java.util.Optional;
 public class SponsorshipService {
     @Autowired
     private SponsorshipDao sponsorshipDao;
+
+    public String setNextId(){
+        List<Sponsorship> sponsorshipList=sponsorshipDao.findAll();
+        int curMaxId=0;
+        for(Sponsorship s:sponsorshipList){
+            curMaxId=Math.max(curMaxId,Integer.parseInt(s.getId()));
+        }
+        int result=curMaxId+1;
+        return result+"";
+    }
 
     public int isExist(String sponsorid,String subjectid){
         Sponsorship checkSponsorship=new Sponsorship();
@@ -64,6 +68,7 @@ public class SponsorshipService {
 
     public void deleteByPK(String sponsorid,String subjectid){
         Sponsorship sponsorship=this.findByPKNo(sponsorid,subjectid);
+        if(sponsorship!=null)
         sponsorshipDao.delete(sponsorship);
     }
 
@@ -73,6 +78,7 @@ public class SponsorshipService {
         int exist=this.isExist(sponsorid,subjectid);
         int success=0;
         if(exist==0){
+            sponsorship.setId(this.setNextId());
             sponsorshipDao.save(sponsorship);
             success=1;
             System.out.println("增加成功");
@@ -86,6 +92,8 @@ public class SponsorshipService {
     public int changeSponsorship(Sponsorship newSponsorship){
         int exist=this.isExist(newSponsorship.getSponsorId(),newSponsorship.getSubjectId());
         if(exist==1){
+            Sponsorship targetSponsorship=this.findByPKNo(newSponsorship.getSponsorId(),newSponsorship.getSubjectId());
+            newSponsorship.setId(targetSponsorship.getId());
             sponsorshipDao.save(newSponsorship);
         }
         return exist;
